+++ packages/smarthr-ui/rollup.esm.config.js
@@ -9,7 +9,6 @@ import { globSync } from 'glob'
 import preserveDirectives from 'rollup-plugin-preserve-directives'
 import renameNodeModules from 'rollup-plugin-rename-node-modules'
 
-
 import packageJson from './package.json' with { type: 'json' }
 
 const peerDependencies = packageJson.peerDependencies
@@ -41,7 +40,7 @@ export default {
     format: 'es',
     sourcemap: true,
     inlineSources: true,
-    dir: 'esm',
+    dir: 'lib',
     preserveModules: true,
     preserveModulesRoot: 'src',
   },
@@ -80,4 +79,13 @@ export default {
       preventAssignment: true,
     }),
   ],
+  onwarn(warning, warn) {
+    // 'use client'がついていると警告が出るが問題ないため、無視する
+    if (warning.code === 'MODULE_LEVEL_DIRECTIVE') {
+      // 他にもルールが追加される可能性があるため
+      // eslint-disable-next-line smarthr/best-practice-for-unnesessary-early-return
+      return
+    }
+    warn(warning)
+  },
 }