+++ packages/smarthr-ui/src/components/DefinitionList/DefinitionListItem.tsx
@@ -3,6 +3,7 @@ import {
   type FC,
   type PropsWithChildren,
   type ReactNode,
+  isValidElement,
   memo,
   useMemo,
 } from 'react'
@@ -12,11 +13,12 @@ import { spacing } from '../../themes'
 import { Stack } from '../Layout'
 import { Text } from '../Text'
 
+type ObjectTermType = {
+  text: ReactNode
+  styleType?: 'blockTitle' | 'subBlockTitle' | 'subSubBlockTitle'
+}
 type AbstractProps = PropsWithChildren<{
-  term: ReactNode
-  termStyleType?: 'blockTitle' | 'subBlockTitle' | 'subSubBlockTitle'
-  /** @deprecated DefinitionList で items を使う時の props です。children を使ってください。 */
-  description?: ReactNode
+  term: ReactNode | ObjectTermType
   fullWidth?: boolean
   maxColumns?: number
 }>
@@ -28,8 +30,8 @@ const classNameGenerator = tv({
       'smarthr-ui-DefinitionListItem shr-border-b-shorthand shr-min-w-[12em] shr-grow shr-border-dotted',
       'contrast-more:shr-border-b-high-contrast',
     ],
-    termEl: 'smarthr-ui-DefinitionListItem-term',
-    descriptionEl:
+    term: 'smarthr-ui-DefinitionListItem-term',
+    description:
       'smarthr-ui-DefinitionListItem-description shr-ms-[initial] shr-min-h-[calc(1em*theme(lineHeight.normal))] shr-pb-0.25',
   },
   variants: {
@@ -42,21 +44,29 @@ const classNameGenerator = tv({
 })
 
 export const DefinitionListItem: FC<Props> = ({
+  term: orgTerm,
+  children,
   maxColumns,
   fullWidth,
-  term,
-  termStyleType = 'subBlockTitle',
-  description,
-  children,
   className,
 }) => {
+  // HINT: ReactNodeとObjectのどちらかを判定
+  // typeofはnullの場合もobject判定されてしまうため念の為falsyで判定
+  // ReactNodeの一部であるReactElementもobjectとして判定されてしまうためisValidElementで判定
+  const term: ObjectTermType =
+    !orgTerm || typeof orgTerm !== 'object' || isValidElement(orgTerm)
+      ? {
+          text: orgTerm as ReactNode,
+        }
+      : (orgTerm as ObjectTermType)
+
   const classNames = useMemo(() => {
-    const { wrapper, termEl, descriptionEl } = classNameGenerator()
+    const cs = classNameGenerator()
 
     return {
-      wrapper: wrapper({ fullWidth, className }),
-      term: termEl(),
-      description: descriptionEl(),
+      wrapper: cs.wrapper({ fullWidth, className }),
+      term: cs.term(),
+      description: cs.description(),
     }
   }, [className, fullWidth])
   const style = useMemo(
@@ -72,19 +82,19 @@ export const DefinitionListItem: FC<Props> = ({
 
   return (
     <Stack gap={0.25} className={classNames.wrapper} style={style}>
-      <DefinitionTerm styleType={termStyleType} className={classNames.term}>
-        {term}
+      <DefinitionTerm styleType={term.styleType} className={classNames.term}>
+        {term.text}
       </DefinitionTerm>
       <Text as="dd" size="M" color="TEXT_BLACK" leading="NORMAL" className={classNames.description}>
-        {children ?? description}
+        {children}
       </Text>
     </Stack>
   )
 }
 
 const DefinitionTerm = memo<
-  PropsWithChildren<{ styleType: Required<Props>['termStyleType']; className: string }>
->(({ styleType, className, children }) => (
+  PropsWithChildren<{ styleType: ObjectTermType['styleType']; className: string }>
+>(({ styleType = 'subBlockTitle', className, children }) => (
   <Text as="dt" leading="TIGHT" styleType={styleType} className={className}>
     {children}
   </Text>