+++ packages/smarthr-ui/src/components/DefinitionList/stories/VRTDefinitionList.stories.tsx
@@ -10,14 +10,17 @@ export default {
     <Stack>
       {['blockTitle', 'subBlockTitle', 'subSubBlockTitle'].map((termStyleType) =>
         [2, 3, 4].map((maxColumns) => (
-          <DefinitionList
-            {...args}
-            termStyleType={termStyleType as any}
-            maxColumns={maxColumns}
-            key={`${termStyleType}-${maxColumns}`}
-          >
+          <DefinitionList {...args} key={`${termStyleType}-${maxColumns}`}>
             {[...Array(10)].map((_, i) => (
-              <DefinitionListItem term={`定義リストアイテム${i + 1}`} fullWidth={i === 4} key={i}>
+              <DefinitionListItem
+                term={{
+                  text: `定義リストアイテム${i + 1}`,
+                  styleType: termStyleType as any,
+                }}
+                maxColumns={maxColumns}
+                fullWidth={i === 4}
+                key={i}
+              >
                 定義リストアイテム説明{i + 1}
               </DefinitionListItem>
             ))}