+++ packages/smarthr-ui/src/components/FormControl/stories/VRTFormControl.stories.tsx
@@ -1,6 +1,7 @@
 import { FaAddressBookIcon } from '../../Icon'
 import { Input } from '../../Input'
 import { Cluster, Stack } from '../../Layout'
+import { StatusLabel } from '../../StatusLabel'
 import { FormControl } from '../FormControl'
 
 import type { Meta, StoryObj } from '@storybook/react-webpack5'
@@ -9,14 +10,14 @@ export default {
   title: 'Components/FormControl/VRT',
   render: (args) => (
     <Stack gap={4}>
-      {[false, true].map((dangerouslyHide) => (
+      {[false, true].map((unrecommendedHide) => (
         <FormControl
           {...args}
           label={{
             ...args.label,
-            dangerouslyHide,
+            unrecommendedHide,
           }}
-          key={dangerouslyHide.toString()}
+          key={unrecommendedHide.toString()}
         />
       ))}
     </Stack>
@@ -27,7 +28,7 @@ export default {
       text: 'フォームコントロール',
       icon: <FaAddressBookIcon />,
     },
-    statusLabelProps: { type: 'grey', children: '任意' },
+    statusLabels: <StatusLabel type="grey">任意</StatusLabel>,
     subActionArea: (
       <Cluster justify="space-between">
         <div>サブアクションエリア（start)</div>