+++ packages/smarthr-ui/src/components/SideNav/stories/VRTSideNav.stories.tsx
@@ -2,7 +2,7 @@ import { type ComponentProps, Fragment } from 'react'
 
 import { Stack } from '../../Layout'
 import { SideNav } from '../SideNav'
-import { SideNavItemButton } from '../SideNavItemButton'
+import { SideNavItemAnchor, SideNavItemButton } from '../SideNavItemButton'
 
 import { _sideNavItems } from './SideNav.stories'
 
@@ -24,16 +24,13 @@ export default {
         <Fragment key={index}>
           {_casse.map((size, i) => (
             <SideNav {...args} key={`${index}-${i}`} size={size} id={id}>
-              {_sideNavItems.map((item) => (
-                <SideNavItemButton
-                  key={item.id}
-                  id={item.id}
-                  current={item.current}
-                  prefix={item.prefix}
-                >
-                  {item.children}
-                </SideNavItemButton>
-              ))}
+              {_sideNavItems.map((item) =>
+                i % 2 === 0 ? (
+                  <SideNavItemButton {...item} key={item.id} />
+                ) : (
+                  <SideNavItemAnchor {...item} key={item.id} href={`#${index}-${i}`} />
+                ),
+              )}
             </SideNav>
           ))}
         </Fragment>