+++ packages/smarthr-ui/src/components/Switch/stories/Switch.stories.tsx
@@ -10,7 +10,7 @@ export default {
     children: 'ラベル',
   },
   argTypes: {
-    dangerouslyLabelHidden: {
+    unrecommendedLabelHidden: {
       control: 'boolean',
     },
     defaultChecked: {
@@ -27,10 +27,10 @@ export default {
 
 export const Playground: StoryObj<typeof Switch> = {}
 
-export const DangerouslyLabelHidden: StoryObj<typeof Switch> = {
-  name: 'dangerouslyLabelHidden',
+export const UnrecommendedLabelHidden: StoryObj<typeof Switch> = {
+  name: 'unrecommendedLabelHidden',
   args: {
-    dangerouslyLabelHidden: true,
+    unrecommendedLabelHidden: true,
   },
 }